Separate the oneloop version from unroll version of EVD
